+++ src/acconeer/exptool/__init__.py
@@ -4,6 +4,7 @@
 from . import a111, utils
 from ._structs import configbase
 from .pg_process import PGProccessDiedException, PGProcess
+from .utils import USBDevice
 
 
 try: