+++ src/acconeer/exptool/_pyusb/__init__.py
@@ -1,0 +1,4 @@
+# Copyright (c) Acconeer AB, 2022
+# All rights reserved
+
+from .pyusbcomm import PyUsbCdc, PyUsbComm